BUG: erroneous error on clicking sign in
Many developers report the following error after following the README and clicking sign in.

```
Error: `headers` was called outside a request scope. Read more: https://nextjs.org/docs/messages/next-dynamic-api-wrong-context
```

This error is erroneous because the signIn function requires a server scope, and the action is defined correctly with 'use server.' Changing this to use the client causes other issues. 

An alternative method to fixing this would be to keep the forms and have the action return the signIn function, as recommended by AI. Unfortunately, this approach results in the actions being out of spec. While it does work on Chrome, it is not guaranteed to be compatible with every browser.

This change has been tested to work as well as the original in desktop and mobile environments in Chrome, Firefox, and Safari. Since many developers rely on this as a starting point, making the development experience as smooth as possible is essential.